Removed mini css plugin and all associated className in components. Installed Tailwind CSS
