+++ client/src/App.tsx
@@ -1,25 +1,22 @@
-import TasksPage from "./tasks/TasksPage";
-import TaskPage from "./tasks/TaskPage";
-import HomePage from "./home/HomePage";
+import TasksPage from "./components/tasks/TasksPage";
+import TaskPage from "./components/tasks/TaskPage";
+import HomePage from "./components/home/HomePage";
 import { BrowserRouter, Routes, Route, NavLink } from "react-router";
 import logo from "./assets/abstract-logo.png"
 
 function App() {
   return (
     <BrowserRouter>
-      <header className="sticky">
-        <span className="logo">
-          <img src={logo} alt="logo" width="49" height="99" />
-        </span>
-        <NavLink to="/" className="button rounded">
-          <span className="icon-home"></span>
+      <header>
+        <img src={logo} alt="logo" width="49" height="99" />
+        <NavLink to="/">
           Home
         </NavLink>
-        <NavLink to="/tasks" className="button rounded">
+        <NavLink to="/tasks">
           Tasks
         </NavLink>
       </header>
-      <div className="container">
+      <div>
         <Routes>
           <Route path="/" element={<HomePage />} />
           <Route path="/tasks" element={<TasksPage />} />
