+++ client/src/components/home/HomePage.tsx
@@ -1,5 +1,5 @@
 function HomePage() {
-    return <h2>Home</h2>
+    return <h1>Home</h1>
 }
 
 export default HomePage;
