+++ client/src/index.css
@@ -1,27 +1,1 @@
-@import "../node_modules/mini.css/dist/mini-default.min.css";
-
-form {
-    min-width: 300px;
-}
-
-html,
-body,
-#root,
-.container,
-.center-page {
-    height: 100%;
-}
-
-.center-page {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-}
-
-header {
-    height: 5.1875rem;
-}
-
-a.button.active {
-    border: 1px solid var(--fore-color);
-}
+@import "tailwindcss";
